+++ module/src/main/java/com/annimon/tgbotsmodule/api/methods/games/SetGameScoreMethod.java
@@ -65,12 +65,12 @@ public class SetGameScoreMethod implements
     }
 
     @Override
-    public Integer getUserId() {
+    public Long getUserId() {
         return method.getUserId();
     }
 
     @Override
-    public SetGameScoreMethod setUserId(@NotNull Integer userId) {
+    public SetGameScoreMethod setUserId(@NotNull Long userId) {
         method.setUserId(userId);
         return this;
     }
