+++ module/src/main/java/com/annimon/tgbotsmodule/api/methods/info/GetUserProfilePhotosMethod.java
@@ -25,12 +25,12 @@ public class GetUserProfilePhotosMethod implements
     }
 
     @Override
-    public Integer getUserId() {
+    public Long getUserId() {
         return method.getUserId();
     }
 
     @Override
-    public GetUserProfilePhotosMethod setUserId(@NotNull Integer userId) {
+    public GetUserProfilePhotosMethod setUserId(@NotNull Long userId) {
         method.setUserId(userId);
         return this;
     }
